// src/app/models/general.ts
export interface General {
    $key: string;
    idEstudiante: string;
    cedula: string;
    grupoSanguineo: string;
    antecedentes: string;
    motivoConsulta: string;
    enfermedadActual: string;
    presion: string;
    temperatura: string;
    saturacionOxigeno: string;
    peso: string;
    talla: string;
    examenFisico: string;
    idCie10: string;
    codigo: string;
    descripcion: string;
    tratamiento: string;
    seguimiento: string;
    observaciones: string;
    estado: string;
    fechaCreacion: Date;
}